backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# PDF Processing
PyMuPDF==1.23.21
pdf2image==1.17.0
pillow==10.2.0

# OCR
pytesseract==0.3.10

# AI/ML (optional, for advanced extraction)
transformers==4.37.2
torch==2.2.0
sentencepiece==0.1.99

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Progress tracking
tqdm==4.66.1

# Image processing
opencv-python-headless==4.9.0.80
numpy==1.26.3
